+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # CHANGELOG
 
+## v0.3.0
+
+Breaking change:
+* On the Query class, `select`, `update`, `insertInto`, `deleteFrom`, and `with` are now static
+  methods, which return a new instance, rather than instance methods.
+* New instance methods `appendSelect`, `appendUpdate`, etc., are provided to replace the old
+  instance method versions of the above functions.
+
 ## v0.2.0
 
+Breaking change:
 * Remove pointless Query::clear() method.
 
 ## v0.1.1